Remove old regio tiles ; move dynamic imports index file one level up
