+++ src/pages/landelijk/besmettelijke-mensen.tsx
@@ -5,7 +5,7 @@ import BarScale from 'components/barScale';
 import Legenda from 'components/legenda';
 import { FCWithLayout } from 'components/layout';
 import { getNationalLayout } from 'components/layout/NationalLayout';
-import { AreaChart } from 'components/tiles/index';
+import { AreaChart } from 'components/index';
 
 import Ziektegolf from 'assets/ziektegolf.svg';
