+++ src/pages/landelijk/verdenkingen-huisartsen.tsx
@@ -4,7 +4,7 @@ import BarScale from 'components/barScale';
 import { ContentHeader } from 'components/layout/Content';
 import { FCWithLayout } from 'components/layout';
 import { getNationalLayout } from 'components/layout/NationalLayout';
-import { LineChart } from 'components/tiles/index';
+import { LineChart } from 'components/index';
 
 import Arts from 'assets/arts.svg';
