+++ src/pages/landelijk/verpleeghuis-besmette-locaties.tsx
@@ -4,7 +4,7 @@ import BarScale from 'components/barScale';
 import { ContentHeader } from 'components/layout/Content';
 import { FCWithLayout } from 'components/layout';
 import { getNationalLayout } from 'components/layout/NationalLayout';
-import { LineChart } from 'components/tiles/index';
+import { LineChart } from 'components/index';
 
 import Locatie from 'assets/locaties.svg';
