+++ src/pages/landelijk/verpleeghuis-positief-geteste-personen.tsx
@@ -4,7 +4,7 @@ import BarScale from 'components/barScale';
 import { FCWithLayout } from 'components/layout';
 import { getNationalLayout } from 'components/layout/NationalLayout';
 import { ContentHeader } from 'components/layout/Content';
-import { LineChart } from 'components/tiles/index';
+import { LineChart } from 'components/index';
 
 import Getest from 'assets/test.svg';
