+++ src/pages/landelijk/ziekenhuis-opnames.tsx
@@ -3,7 +3,7 @@ import useSWR from 'swr';
 import BarScale from 'components/barScale';
 import { FCWithLayout } from 'components/layout';
 import { getNationalLayout } from 'components/layout/NationalLayout';
-import { LineChart } from 'components/tiles/index';
+import { LineChart } from 'components/index';
 import { ContentHeader } from 'components/layout/Content';
 
 import Ziekenhuis from 'assets/ziekenhuis.svg';
